+++ sqitch/deploy/function_authenticate.sql
@@ -15,7 +15,7 @@ CREATE FUNCTION maevsi.authenticate(
 ) RETURNS maevsi.jwt AS $$
 DECLARE
     "_jwt_id" UUID := maevsi.uuid_generate_v1mc();
-    "_jwt_exp" BIGINT := EXTRACT(EPOCH FROM ((SELECT date_trunc('second', NOW()::TIMESTAMP)) + COALESCE(current_setting('maevsi.jwt_expiry_duration'), '1 day')::INTERVAL));
+    "_jwt_exp" BIGINT := EXTRACT(EPOCH FROM ((SELECT date_trunc('second', NOW()::TIMESTAMP)) + COALESCE(current_setting('maevsi.jwt_expiry_duration', true), '1 day')::INTERVAL));
     "_jwt" maevsi.jwt;
 BEGIN
   IF ("username" = '' AND "password" = '') THEN