+++ sqitch/deploy/function_redeem.sql
@@ -22,11 +22,11 @@ BEGIN
     "_jwt_id" := current_setting('jwt.claims.id', true)::UUID;
     "_jwt" := (
         "_jwt_id",
-        current_setting('jwt.claims.exp', true)::BIGINT,
         current_setting('jwt.claims.role', true)::TEXT,
         current_setting('jwt.claims.account_id', true)::INTEGER,
         current_setting('jwt.claims.username', true)::TEXT,
         (SELECT ARRAY(SELECT DISTINCT UNNEST(maevsi.invite_claim_array() || "invitation_code") ORDER BY 1))
+        current_setting('jwt.claims.exp', true)::BIGINT,
     )::maevsi.jwt;
 
     UPDATE maevsi_private.jwt